Initial3

+++ app/src/main/java/com/example/appdevfinalproject/CartItem.java
@@ -2,12 +2,12 @@ package com.example.appdevfinalproject;
 
 public class CartItem {
     private String productName;
-    private double productPrice;
+    private int productPrice;
     private int productImage;
     private String username; // Added field for username
 
 
-    public CartItem(String username, String productName, double productPrice, int productImage) {
+    public CartItem(String username, String productName, int productPrice, int productImage) {
         this.username = username;
         this.productName = productName;
         this.productPrice = productPrice;
@@ -19,7 +19,7 @@ public class CartItem {
         return productName;
     }
 
-    public double getProductPrice() {
+    public int getProductPrice() {
         return productPrice;
     }
